add detail/json_ref.hpp
