+++ Makefile
@@ -21,7 +21,8 @@ SRCS = ${SRCDIR}/json.hpp \
 			 ${SRCDIR}/detail/parsing/output_adapters.hpp \
 			 ${SRCDIR}/detail/parsing/binary_reader.hpp \
 			 ${SRCDIR}/detail/parsing/binary_writer.hpp \
-			 ${SRCDIR}/detail/serializer.hpp
+			 ${SRCDIR}/detail/serializer.hpp \
+			 ${SRCDIR}/detail/json_ref.hpp
 
 # main target
 all:
